pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Michael Peng", email = "michael@mcmoo.org"},
]
dependencies = [
    "certifi==2021.10.8",
    "charset-normalizer==2.0.12",
    "idna==3.3",
    "python-dotenv==0.20.0",
    "requests==2.27.1",
    "urllib3==1.26.9",
    "Pillow~=10.0.0",
    "websocket-client~=1.3.2",
    "colorama==0.4.4",
    "websockets>=11.0.3",
    "aioconsole~=0.4.1",
    "beautifulsoup4>=4.12.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
